Refactor package to zeo
