--- src/pyzeo/area_volume/__init__.py
@@ -1,1 +1,0 @@
-from pyzeo.extension import volume, surface_area
