--- src/pyzeo/netstorage/__init__.py
@@ -1,1 +1,0 @@
-from pyzeo.extension import AtomNetwork
