--- src/pyzeo/psd/__init__.py
@@ -1,1 +1,0 @@
-from pyzeo.extension import calc_pore_size_distribution
