+++ src/zeo/cluster/__init__.py
@@ -1,4 +1,4 @@
-from pyzeo.extension import (
+from zeo.extension import (
     pruned_highaccuracy_voronoi_network,
     get_nearest_largest_diameter_highaccuracy_vornode,
 )
