+++ src/zeo/cycle/__init__.py
@@ -1,4 +1,4 @@
-from pyzeo.extension import (
+from zeo.extension import (
     compute_centroid_4cycles,
     compute_face_centers
 )
