+++ src/zeo/netstorage/__init__.py
@@ -1,0 +1,1 @@
+from zeo.extension import AtomNetwork
