+++ src/zeo/psd/__init__.py
@@ -1,0 +1,1 @@
+from zeo.extension import calc_pore_size_distribution
